Clean up Data Subpackage
* Move routines to routines.py
* Move analyze_beammap to rfsocinterface.analysis
* Removed old ProcessedData and MapData classes and duplicated functions
